DEV: Reorganize drafts controller
