+++ app/assets/javascripts/discourse/app/lib/text.js
@@ -110,3 +110,55 @@ export function emojiUrlFor(code) {
     return buildEmojiUrl(code, opts);
   }
 }
+
+function encode(str) {
+  return str.replaceAll("<", "&lt;").replaceAll(">", "&gt;");
+}
+
+function traverse(element, callback) {
+  if (callback(element)) {
+    for (let i = 0; i < element.childNodes.length; ++i) {
+      traverse(element.childNodes[i], callback);
+    }
+  }
+}
+
+export function excerpt(cooked, length) {
+  let result = "";
+  let resultLength = 0;
+
+  const div = document.createElement("div");
+  div.innerHTML = cooked;
+  traverse(div, (element) => {
+    if (resultLength >= length) {
+      return;
+    }
+
+    if (element.nodeType === Node.TEXT_NODE) {
+      if (resultLength + element.textContent.length > length) {
+        const text = element.textContent.substr(0, length - resultLength);
+        result += encode(text);
+        result += "&hellip;";
+        resultLength += text.length;
+      } else {
+        result += encode(element.textContent);
+        resultLength += element.textContent.length;
+      }
+    } else if (element.tagName === "A") {
+      element.innerHTML = element.innerText;
+      result += element.outerHTML;
+      resultLength += element.innerText.length;
+    } else if (element.tagName === "IMG") {
+      if (element.classList.contains("emoji")) {
+        result += element.outerHTML;
+      } else {
+        result += "[image]";
+        resultLength += "[image]".length;
+      }
+    } else {
+      return true;
+    }
+  });
+
+  return result;
+}
